New version

+++ src/remote/delete.js
@@ -1,4 +1,7 @@
 const path = require('path');
+const PERMISSION_DENIED = require('./error_messages').PERMISSION_DENIED;
+const NO_SUCH_FILE = require('./error_messages').NO_SUCH_FILE;
+const tracer = require('../tracer');
 
 module.exports = (list, remoteRoot, deployment) => {
     return new Promise((resolve, reject) => {
@@ -9,26 +12,38 @@ module.exports = (list, remoteRoot, deployment) => {
 function deleteFiles(list, remoteRoot, deployment, resolve, reject) {
     let fileToDelete = list.pop();
     if(fileToDelete){
-        process.stdout.write("\n"+'Removing file '+fileToDelete.name);
+        tracer.processStart('Removing file '+fileToDelete.name)
         deployment.delete(remoteRoot + fileToDelete.name).then(() => {
-            process.stdout.write('...OK');
+            tracer.processSuccess();
             let dirname = path.dirname(remoteRoot + fileToDelete.name);
             deployment.list(dirname).then(files => {
                 if(files.length){
                     deleteFiles(list, remoteRoot, deployment, resolve, reject);
                 } else {
+                    tracer.processStart('Removing directory '+dirname);
                     deployment.rmdir(dirname).then(() => {
-                        process.stdout.write("\n"+'Removing directory '+dirname);
+                        tracer.processSuccess();
                         deleteFiles(list, remoteRoot, deployment, resolve, reject);
+                    }).catch(e => {
+                        if(e.message == PERMISSION_DENIED){
+                            tracer.processError(PERMISSION_DENIED);
+                            deleteFiles(list, remoteRoot, deployment, resolve, reject);
+                        } else {
+                            throw e;
+                        }
                     })
                 }
-            }).catch(e => {
-                process.stdout.write("\n"+'Failed to access directory '+dirname);
-                deleteFiles(list, remoteRoot, deployment, resolve, reject);
             })
         }).catch(e => {
-            process.stdout.write("\n"+'Failed to delete file '+remoteRoot + fileToDelete.name+' with message: '+e.message);
-            deleteFiles(list, remoteRoot, deployment, resolve, reject);
+            if(e.message == NO_SUCH_FILE){
+                tracer.processError(NO_SUCH_FILE);
+                deleteFiles(list, remoteRoot, deployment, resolve, reject);
+            } else if (e.message == PERMISSION_DENIED){
+                tracer.processError(PERMISSION_DENIED);
+                deleteFiles(list, remoteRoot, deployment, resolve, reject);
+            } else {
+                throw e;
+            }
         });
     } else {
         resolve();
